Quite a few layout changes:     * taxonomy support     * intro and tag list in list headers (todo: add for groups)     * sorting by weight     * layout in rows in order, no more need to balance columns     * image caching in content folder
The last one is a big change and probably a bit of a hack. Hugo allows
for image resizing but requires the original to be present at every build,
even though the resized image is cached.
This is inefficient in my case as I am looking at about 50G of immutable
assets. Checking out all of them for every build is silly and wasteful.

The new system allows the admin to move the resized pictures from
resources/_gen/images/<album> to content/<album>. Afterwards the originals
can be removed. (I check them into a submodule for safekeeping.)

The templates get the list of images only from the resource descriptions
in _index. For each image they first look for the orignal in the asset folder.
If no org is found, the already resized version from the content folder
is taken instead.

{TODO: document this in the readme}