Adopt & document new hugo author name/email param
Also - display the author email, if specified, in the meta author tag
alongside name.  I'm not sure there's a standard way to do this, but it
doesn't seem to be against spec and presumably if people are putting
info in this field they're kinda expecting it to show somewhere other
than just rss.

Also - provide an example of how the author name/email can be album
specific - it's in the fish album.